+++ src/games/chess/controller.rs
@@ -57,7 +57,7 @@ impl<const T: usize> MoveController<T> for ChessMoveController<T> {
                             "{}{}{}",
                             board.encode_position(from),
                             board.encode_position(action.to),
-                            board.game.pieces[action.piece_type].format_info(board, action.info)
+                            board.game.pieces[action.piece_type as usize].format_info(board, action.info)
                         ),
                         None => "----".to_string()
                     }
